requirements-production.txt
astral==2.2
alabaster==0.7.12
amqp==2.6.0
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
Beaker==1.11.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cryptography==3.0
decorator==4.4.2
Django==3.2.4
django-admin-hstore-widget==1.1.0
django-constance==2.8.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-proxy==1.2.1
django-ranged-response==0.2.0
django-redis==5.0.0
django-simple-captcha==0.5.14
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-gis==0.17
Fiona==1.8.17
future==0.18.2
GDAL==2.4.0
geographiclib==1.50
geopy==2.0.0
gevent==20.9.0
gitdb==4.0.5
GitPython==3.1.7
greenlet==0.4.17
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
josepy==1.4.0
jsonschema==3.2.0
kombu==4.6.11
lazy-object-proxy
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mozilla-django-oidc==1.2.4
munch==2.5.0
numpy==1.19.1
oauth2client==4.1.3
ogr==0.23.0
oic==0.10.0.0
paramiko==2.7.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
spatialite==0.0.3
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.8
PyGithub==1.51
Pygments==2.6.1
pyjwkest==1.4.2
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyproj==2.6.1.post1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-gitlab==2.4.0
pytz==2020.1
django-ckeditor==6.7.2
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
responses==0.10.15
rsa==4.6
Rtree==0.9.7
Shapely==1.7.1
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
testfixtures==6.14.1
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.1
ummalqura==2.0.1
unicodecsv==0.14.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
Whoosh==2.7.4
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
django-json-widget==1.0.1
django-verbatim==0.1
django-model-utils==4.0.0
py3rijndael==0.3.3
python-dotenv
django-ordered-model==3.7.4
django-richtextfield==1.6.2
django-ckeditor-5==0.2.13
django-photologue==3.17